view & controller cleanups
