+++ src/main/java/david/buzas/whattoeat/ui/controllers/Controller.java
@@ -3,11 +3,11 @@ package david.buzas.whattoeat.ui.controllers;
 import david.buzas.whattoeat.ui.components.ErrorAlert;
 
 public class Controller {
-    protected void showError(String message) {
+    protected void showErrorMessage(String message) {
         new ErrorAlert(message).showAndWait();
     }
 
-    private void showError(Exception exception) {
-        this.showError(exception.getMessage());
+    protected void showError(Exception exception) {
+        this.showErrorMessage(exception.getMessage());
     }
 }
