--- src/main/java/david/buzas/whattoeat/ui/controllers/MainController.java
@@ -1,6 +1,0 @@
-package david.buzas.whattoeat.ui.controllers;
-
-
-public class MainController {
-
-}
